Unifica controle da ARSession
